Rename lib/logger.rb to lib/narou_logger.rb in order to avoid conflicts.
