+++ lib/narou_logger.rb
