+++ spec/convert_spec.rb
@@ -7,7 +7,7 @@
 Encoding.default_external = Encoding::UTF_8
 
 require_relative "../lib/commandline"
-require_relative "../lib/logger"
+require_relative "../lib/narou_logger"
 
 AUTHOR = "whiteleaf"
 $debug = File.exist?(File.expand_path(File.join(File.dirname(__FILE__), "../debug")))
