+++ spec/generator/convert_spec_gen.rb
@@ -47,7 +47,7 @@ __END__
 Encoding.default_external = Encoding::UTF_8
 
 require_relative "../lib/commandline"
-require_relative "../lib/logger"
+require_relative "../lib/narou_logger"
 
 AUTHOR = "whiteleaf"
 $debug = File.exist?(File.expand_path(File.join(File.dirname(__FILE__), "../debug")))
